refactor: move the DataFrame seam into panel.util
Not re-exported from panel.util, since util already exports a pandas
specific is_dataframe from checks and the two names would collide.
Importing from panel.util.dataframe keeps the distinction visible.
